Add test that verifies file:line
We want to verify the line number matches what we expect, especially
since the unsafe version works slightly differently which can lead to
slightly different line numbers.

Hardcoding line numbers makes tests difficult to maintain, so go with
an approach where we look for a marker comment and find that marker to
represent the line number we expect.

The line numbers aren't what we always expect -- I think this is due to
FuncForPC being used, will follow-up.