line test: scan the source for comments
Currently, we use "// trace line" comments as markers for line numbers
that we want to see in the error trace of a function.
We do this with simple string parsing.
This approach is fragile and limited:
it only works for this exact subtest pattern,
and it allows for only one marker per subtest.

This switches to using go/scanner to scan the source file for tokens,
grouping zero or more trace line numbers under group names.
Group names are identified by `// @group` comments
following a string literal.
All `// @trace` comments following that comment
until the next group or EOF are considered part of that group.

Right now,
this is just a more accurate way to implement what we're already doing.
In the future this can be expanded to allow for custom group names,
e.g.

    // @group:"foo"