Fix pointer arithmetic caught by `-race`
When doing pointer arithmetic, the conversion from `unsafe.Pointer` to
`uintptr`, and then back, must be within a single expression.

From https://pkg.go.dev/unsafe#Pointer
> // INVALID: uintptr cannot be stored in variable
> // before conversion back to Pointer.

This is caught by `-race`, so enable it in the Makefile.

Note that `-race` is only supported for amd64/arm64, so disable
it on other architectures using the GitHub action.