cmd: Add an errtrace command
Adds a new errtrace binary that instruments existing Go code
with errtrace.Wrap calls.

This operates by searching for functions that return errors,
and inside those functions, wrapping all returns with `errtrace.Wrap`.

To allow for future usage with -toolexec or similar,
the code edits all ensure that line numbers do not change.
We may want to reconsider this behavior in the future.

Pending:

- More testing with more corner cases
- Detect overlapping edits: this should be considered a bug
- Handling of statements like `return foo()`
  where `foo()` returns `x, err` -- this will currently panic.
  This needs helper `Wrap{N}` functions
  parameterized over the non-error return types,
  and that still won't handle multiple error returns
  like the rest of the program. That should be fine.
- Handling of package naming conflict:
  The code has room for using a name other than "errtrace"
  for the package, but the conflict isn't currently detected.
  This probably needs making two passes over the file
  to find a unique identifier.