Detect and panic on overlapping edits
The editing scheme might not be our best option here.
This model cannot handle

    return func() error {
        return errors.New("foo")
    }()

Because the outer closure overlaps with the inner.

I think the editing scheme needs to separate open and close of wrapping
so that "add errtrace.Wrap(" and "add )" are separate edits,
that do not overlap with the inside of the return expression.