pc: Replace amd64 unsafe implementation with assembly, drop 386 unsafe implementation (#12)
The current amd64 implementation gets the caller's PC based on the argument of
the first address. This approach can be implemented fully in Go (so it can be
inlined, and take advantage of register calling conventions). However, it
assumes a specific layout which has a couple of issues:

 * When using generics, e.g., `Wrap2[T any](T, error)(T, error)`, the address
   of the first argument can't be used as-is, as there's an additional hidden
   dictionary argument, see:
   https://github.com/golang/proposal/blob/master/design/generics-implementation-dictionaries-go1.18.md

 * It technically is a violation according to the [register-based calling
   convention](https://go.googlesource.com/proposal/+/refs/changes/78/248178/1/design/40724-register-calling.md):

> Code that performs unsafe.Pointer arithmetic on pointers to arguments in
> order to observe the contents of the stack. This is a violation of the
> unsafe.Pointer rules today.

We can an assembler implementation instead, which relies on frame pointers.
Callers copy the current `SP` into `BP`, which callees are expected to push
to the stack before modifying. This ends up as a linked-list, see:
https://blog.felixge.de/reducing-gos-execution-tracer-overhead-with-frame-pointer-unwinding/

Specifically,
> The process starts by reading the value of the r29 (arm64) or rbp (amd64)
> register (aka base pointer).  This takes us to the frame pointer of frame 2
> One word (8 bytes) above this value sits the return address of the caller,
> which is the first program counter that we need to collect

The amd64 implementation is effectively: `return *(BP + 8)`.

Frame pointers are not enbaled on 386, so this change removes the 386 unsafe
path for now.