pc: Reference getfp and make arm64 consistent with getfp
To make the asm implementations easier to maintain, use the `getfp`
functions in the runtime as a reference for the FP register we use.

The return address always sits one word above the frame pointer,
https://github.com/golang/go/blob/dc094f9c9613a2a8ed24ace1b5416170aa3a334b/src/runtime/trace2stack.go#L206

This change only adds comments to amd64, and updates the register+offset
used for arm64 for consistency with amd64.