Format: Support multi-errors
New
---

Adds support for formatting multi-errors.
Singular errors are formatted the same as they are today.
However, if an error chain ends hits a multi-error
we'll turn that into a tree structure.

Everything up until the multi-error
is considered part of the same trace node.
The individual children of the multi-error
form the children of that node.

The tree structure is printed in ASCII
using output like the following

```
foo
	foo.go:42
bar
	bar.go:24
|
+- baz
|  	baz.go:24
|  qux
|  	qux.go:48
|  |
|  +- quux
|     	quux.go:24
|     quuz
|     	quuz.go:48
|
+- corge
   	corge.go:24
   grault
   	grault.go:48
```

Changed
-------

We now iterate through the error structure
by manually casting it to the `Unwrap()`-based interfaces
similar to `errors.As`.
https://cs.opensource.google/go/go/+/refs/tags/go1.21.4:src/errors/wrap.go;l=118-123

This allows us to properly handle the logic where
errors until the first multi-error are part of the same trace.

Removed
-------

Because this changes how we iterate through errors,
I've removed the `Frame` type and the `Frames` function
as those are no longer able to represent the actual structure.