pc: Remove GetCaller wrapper for asm
The `GetCaller` wrapper is only used to call the assembly function
but to avoid it adding an unnecessary frame, we use `go:inline`.
This does not work when inlining is disabled, so instead remove the
wrapper and have the assembly function be exported directly.

Update the test target to verify that the tests pass without inlining
and optimizations, which fails without this fix.