cmd: Support package patterns
Adds support to the errtrace command
for recognizing import path patterns.
This makes the common case for command usage:

    errtrace -w ./...

This is an optional addition to the existing file-based behavior.
If the provided argument is an existing file, it will be used as-is.
Otherwise, the pattern will be expanded with`go list`.

If a pattern is provided, all matching Go files are in-scope:
source, test, and external tests.

To cover the case of build-tagged files, we also need a `-tags` flag.
I considered a version where we use `go list` to find directories,
and then transformed all `.go` files in that directory,
but that still breaks when all files in a directory are build-tagged.
If we're going to need a `-tags` argument anyway,
then it's best to match expectations folks have with tags.

As an added benefit:
We'll run `go list` with the `errtrace` tag set.
This makes it possible for entire files to opt-out
by adding `//go:build !errtrace` on top.

Resolves #73