cmd: Print a message when reading from stdin implicitly
One thing that can be confusing about the UX of the CLI is that
when you call it without any arguments,
it's reading from stdin but you may not know that.
You may assume it's just taking forever to start up.

This changes it to print a message to stderr
if it's reading from stdin because no arguments were given.