cmd: Instrument with errtrace
Instruments the errtrace cmd with itself.
Example output:

```
% ./errtrace -w -
stdin:can't use -w with stdin

main.(*mainCmd).readFile
        [..]/errtrace/cmd/errtrace/main.go:571
main.(*mainCmd).processFile
        [..]/errtrace/cmd/errtrace/main.go:351
```

Includes a lint check to ensure that the command remains instrumented.

Instrumentation is opt-in in the Makefile.
We don't want to instrument the errtrace library or its dependencies
because that will introduce a circular dependency.