cmd/errtrace: Best-effort detection of tty stdin using char device
Follow-up to #78 and #82.

When stdin is implicitly used, we wait for 200ms for input before
printing a message to stderr to notify the user that we're reading
from stdin to avoid confusion. This added some complexity in the read
path, as well as the testing because of the background goroutine that
waits for a timeout before printing to stderr.

This is an alternative approach that does a best-effort detection of a
TTY by checking if stdin is a character device.