cmd/errtrace: Add initial toolexec mode
Use TOOLEXEC_IMPORTPATH to determine if we're running in toolexec mode.

In toolexec mode, all go tool executions are intercepted, though only
compile commands interacting with .go files are modified.

Only .go files that import errtrace are rewritten to a temporary file
and passed to compile instead of the original path.