cmd/errtrace: Add no-wrapn option to avoid WrapN usage (#95)
This unblocks #91, as `go:linkname` can't be used to link to generic
functions, which `WrapN` functions rely on.

Rewrites use temporary variables to hold the returns from a function
call before returning, wrapping the error with `Wrap`:
```
return SomeFuncCall()

{ r1, ..., rN := SomeFuncCall(); return r1, ..., errtrace.Wrap(rN) }
```

A new scope is used to ensure variables don't clash, avoiding the need
to track what variables are in scope.

Most of our golden files aren't impacted by this change, so we run tests
with no options, and with no-wrapn. For tests that are impacted, they
specify which options are required to run the test.