Add test for shadowing named return value
We go to some lengths to make sure that
we're tracking the correct named return value in assignments,
but we don't have a test for it.

This adds a test that verifies that if a named return value is shadowed,
the shadowed value is not wrapped if it's not otherwise in scope for
wrapping.