toolexec: Add unsafe rewrite support (#128)
Fixes #91.

Add ability to rewrite packages that don't import errtrace to be
rewritten to wrap errors automatically. This would normally cause
errors when trying to build the package as we're adding a new
dependency to the package, which toolexec doesn't support:
https://github.com/golang/go/issues/35204

To workaround this, we use the hack mentioned in the above issue
using `go:linkname` to "import" errtrace functionality in a way
that doesn't break the package build.

Using `go:linkname` requires importing `unsafe`, to the rewriter
is updated to:
 * import the "unsafe" package instead of errtrace, which can be added
   as a new dependency without impacting the build as it's stdlib.
 * Add `go:linkname` to import errtrace functions into a package.
 * Rewrite to use the `go:linkname` errtrace identifiers instead of
   the errtrace package.

`go:linkname` is not compatible with generic functions like `WrapN` so
disable `WrapN` when using toolexec. `WrapN` is for `errtrace -w` where
source files are directly modified, and formatted with `gofmt` where
`WrapN` minimizes rewriting. This isn't necessary for toolexec mode.

Unsafe mode still requires one package that's part of the final linked
object references `errtrace` somewhere in the binary -- otherwise, the
build fails. This seems like a reasonable limitation since users need to
manage the dependency.